repo refactor done, testing application and moving onto testing refactoring to accommodate this refactoring
