+++ purch/api/routers/__init__.py
@@ -1,0 +1,3 @@
+from purch.api.routers.auth import router as auth_router
+from purch.api.routers.user import router as user_router
+from purch.api.routers.finance import router as finance_router
