+++ purch/main.py
@@ -4,9 +4,11 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from purch.taskiq import broker
-from purch.user.router import router as user_router
-from purch.auth.router import router as auth_router
-from purch.finance.router import router as finance_router
+from purch.api.routers import (
+    user_router, 
+    auth_router, 
+    finance_router,
+)
 from purch.utils.project_version import version
 from purch.api.startup import init_db
 from purch.common.logger import get_logger
